+++ src/main/resources/templates/demo.html
@@ -103,7 +103,6 @@ Headache, new loss of taste. </textarea>
                 <p id="output-msg">JSON response for <a href="https://brat.nlplab.org/embed.html" target="_blank">brat</a>:</p>
                 <div id="output" style="width: 100%;">
                     <pre id="output-json"></pre>
-
                     <p> Visualization by <a href="https://brat.nlplab.org/embed.html" target="_blank">brat</a>: </p>
                     <div id="fig_bratvis">
